executor: Move the existing ratelimit executor to its own package
In order to make place for additional executor types, the existing
ratelimit one is placed under the `ratelimit` package.
