+++ pkg/executor/ratelimit/pool.go
@@ -17,7 +17,7 @@
  *
  */
 
-package executor
+package ratelimit
 
 import (
 	"sync"
