+++ pkg/virt-handler/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
         "//pkg/container-disk:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/ephemeral-disk-utils:go_default_library",
-        "//pkg/executor:go_default_library",
+        "//pkg/executor/ratelimit:go_default_library",
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/host-disk:go_default_library",
         "//pkg/hotplug-disk:go_default_library",
